Fixed the next issue
The Problem: when there are two paths between two objects, the path traversed last
gets used for next which sometimes doesn't make sense.
For example consider this scenario

Path 1: bruce -> resides_in -> Gotham <- resides_in <- John Wayne
in this case:
bruce._relations['resides_in'][0]._next._relations['resides_in'][0]._next.name
Gets us John Wayne

Path 2: bruce -> teaches -> oop <- studies <- John Wayne -> resides in -> Gotham
In this case for:
bruce._relations['resides_in'][0]._next._relations['resides_in'][0]._next
We get Gotham Area object which has no name attribute.
This is in-correct and occurs because we're reusing the relationship
"John Wayne resides in Gotham", but since this time the parent object is John Wayne
Because we're coming from the other side, Gotham is set as the _next object, overwriting
The previous _next object Which was John Wayne

Solution:
When we fetch a relationship object from existing store, we need to check if it's _next
does not match our current parent_object. If it does, we should create a new
relationship object. We also store the relationships with their parent object
IDs into the relations_added dict so we can fetch the correct one accordingly
we'll be fetching according to the relationship and parent_object.